Refactor LocationManager class
