+++ Tella/Utils/Extensions/AVCaptureMovieFileOutputExtension.swift
@@ -19,19 +19,8 @@ extension AVCaptureMovieFileOutput {
         locationMetadataItem.keySpace = .quickTimeMetadata
         locationMetadataItem.key = AVMetadataKey.quickTimeMetadataKeyLocationISO6709 as (NSCopying & NSObjectProtocol)?
         
-        let locationString = convertToISO6709(location: location)
+        let locationString = location.convertToISO6709()
         locationMetadataItem.value = locationString as (NSCopying & NSObjectProtocol)?
         self.metadata = [locationMetadataItem]
-    }
-    
-    func convertToISO6709(location: CLLocation) -> String {
-        let latitude = location.coordinate.latitude
-        let longitude = location.coordinate.longitude
-        
-        let latitudeString = String(format: "%+08.4f", latitude)
-        let longitudeString = String(format: "%+09.4f", longitude)
-        
-        let iso6709String = "\(latitudeString)\(longitudeString)/"
-        return iso6709String
     }
 }
